refactor: allow for splitting tests to different files
